% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{propolisSampleList}
\alias{propolisSampleList}
\title{Brazilian propolis from different harvest seasons and agroecological regions (sample list)}
\format{
An object of class \code{"list"}
}
\usage{
propolisSampleList
}
\description{
Propolis, or bee glue, is a sticky dark-colored substance produced
from collected buds or plant exudates (resin) by bees (Apis mellifera L.).
The resin is masticated, salivary enzymes are added, and the partially
digested material is mixed with beeswax and used in the hive to seal walls,
strengthen comb borders and embalm dead invaders (Wollenweber et al., 1990).
The propolis samples are from NMR data and were collected in autumn (AU),
winter (WI), spring (SP) and summer (SM) of 2010 from Apis mellifera hives
located in southern Brazil (Santa Catarina State). A total of 59 samples
were collected, distributed by season as follows: SM - 16 samples, AU and
SP - 15 samples each, and WI - 13 samples. Three agroecological regions were
defined for the different apiaries and distributed as follows: Highlands -
12 samples, Plain - 11 samples and Plateau - 36 samples.
}
\references{
E. Wollenweber, B. M. Hausen, and W. Greenaway. Phenolic
constituents and sensitizing properties of propolis, poplar balsam and
balsam of Peru. Bulletin de Groupe Polyphenol, 15:112-120, 1990.

M. Maraschin, A. Somensi-Zeggio, S. K. Oliveira, S. Kuhnen,
M. M. Tomazzoli, A. C. M. Zeri, R. Carreira, and M. Rocha.
A machine learning and chemometrics assisted interpretation of spectroscopic
data - a NMR-based metabolomics platform for the assessment of Brazilian
propolis. 2012.
}
\keyword{datasets}
