% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spinalCord}
\alias{spinalCord}
\title{Mouse spinal cord LC-MS data}
\format{
An object of class \code{"list"}
}
\usage{
spinalCord
}
\description{
This dataset consists of 12 LC-MS spectra samples in netCDF
format from mouse spinal cord, divided into two groups: wild type and
knockout. The data was obtained from MetaboAnalyst and originates from a
study describing a general strategy for identifying endogenous substrates of
enzymes by untargeted LC-MS analysis of tissue metabolomes from wild-type and
enzyme-inactivated organisms.
}
\references{
A. Saghatelian, S.A. Trauger, E.J. Want, E.G. Hawkins,
G. Siuzdak, and B.F. Cravatt. Assignment of endogenous substrates to enzymes
by global metabolite profiling. Biochemistry, 43:14332-14339, 2004.
}
\keyword{datasets}
