% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cassavaPPD}
\alias{cassavaPPD}
\title{Cassava Postharvest Physiological Deterioration}
\format{
An object of class \code{"list"}
}
\usage{
cassavaPPD
}
\description{
Cassava is a root well known and widely cultivated in tropical and 
subtropical regions for its starchy tuberous root, which is a great 
source of carbohydrates. It also has a great variety of applications, 
like animal feeding, culinary or alcoholic beverages. In some countries, 
cassava has also been tested as an ethanol biofuel feedstock.
}
\references{
Uarrota et al. (2014) Metabolomics combined with chemometric tools 
(pca, hca, pls-da and svm) for screening cassava (manihot esculenta crantz) 
roots during postharvest physiological deterioration. Food Chemistry 161:67-78
}
\keyword{datasets}
